backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# ML & Data Science
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Deep Learning
torch==2.1.1
transformers==4.36.2

# MLflow
mlflow==2.8.1

# SHAP for explainability
shap==0.43.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Telegram
python-telegram-bot==20.7

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
